+++ CHANGELOG.md
@@ -14,6 +14,9 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Bugfix 🐛:
+- Fixed the reminder still firing when the weight was already logged that day
+
 
 ## [1.4.0] - 2026-09-09